dev update - enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5938 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ branches/dev-three/src/org/hsqldb/util/TableSorter.java
@@ -534,7 +534,7 @@ public class TableSorter extends AbstractTableModel {
 
             // In a compound sort, make each succesive triangle 20%
             // smaller than the previous one.
-            int dx = (int) (size / 2 * Math.pow(0.8, priority));
+            int dx = (int) ((double) size / 2 * Math.pow(0.8, priority));
             int dy = descending ? dx
                                 : -dx;
